feat(calendar): refactor calendar data handling by moving MONTHS, DAYS, and EVENT_TYPES to data file
